Revert zlib to 1.2.3 for easier validation with C++ version of RfgTools
The C++ version of RfgTools used zlib 1.2.3 since the game uses that version. I'm in the process of porting the C++ version to beef. While newer versions of zlib should be compatible, the compressed output isn't identical (possibly using a newer or improved compression algorithm). By using the same version I can just do a plain binary comparison instead of needing more thorough checks to make sure it's just the compression size that differs. The C++ is proven to work reliably, so as long as the port behaves identically it's sufficient. One day it can be upgraded to newer versions of zlib, but for now this old version gets the job done.